restructure: seperate avatar related components
