+++ homeworks/usmonbek-rustamov/18-react-ui/src/components/Avatar/Avatar.tsx
@@ -1,4 +1,4 @@
-import {AvatarData, CamelCase} from '../common/types';
+import {AvatarData, CamelCase} from '../../common/types';
 import refreshImg from '../assets/refresh.svg';
 
 type AvatarProps = {
